Move IamSlice to own file
