+++ pyam/core.py
@@ -23,6 +23,7 @@ except ImportError:
 
 from tempfile import TemporaryDirectory
 
+from pyam.slice import IamSlice
 from pyam.filter import filter_by_time_domain, filter_by_year, filter_by_dt_arg
 
 try:
@@ -89,64 +90,6 @@ from pyam._debiasing import _compute_bias
 from pyam.logging import raise_data_error
 
 logger = logging.getLogger(__name__)
-
-
-class IamSlice(pd.Series):
-    @property
-    def _constructor(self):
-        return IamSlice
-
-    _internal_names = pd.Series._internal_names + ["_iamcache"]
-    _internal_names_set = set(_internal_names)
-
-    def __init__(self, data=None, index=None, **kwargs):
-        super().__init__(data, index, **kwargs)
-        self._iamcache = dict()
-
-    def __dir__(self):
-        return self.dimensions + super().__dir__()
-
-    def __getattr__(self, attr):
-        ret = object.__getattribute__(self, "_iamcache").get(attr)
-        if ret is not None:
-            return ret.tolist() if attr != "time" else ret
-
-        if attr in self.dimensions:
-            ret = self._iamcache[attr] = self.index[self].unique(level=attr)
-            return ret.tolist() if attr != "time" else ret
-
-        return super().__getattr__(attr)
-
-    def __len__(self):
-        return self.sum()
-
-    @property
-    def dimensions(self):
-        return self.index.names
-
-    def __repr__(self):
-        return self.info() + "\n\n" + super().__repr__()
-
-    def info(self, n=80):
-        """Print a summary of the represented index dimensions
-
-        Parameters
-        ----------
-        n : int
-            The maximum line length
-        """
-        # concatenate list of index dimensions and levels
-        info = f"{type(self)}\nIndex dimensions:\n"
-        c1 = max([len(i) for i in self.dimensions]) + 1
-        c2 = n - c1 - 5
-        info += "\n".join(
-            [
-                f" * {i:{c1}}: {print_list(getattr(self, i), c2)}"
-                for i in self.dimensions
-            ]
-        )
-
-        return info
 
 
 class IamDataFrame(object):
